+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ==================
 
   * Bottom bar on hover style fixed
+  * Fixed bug where more than the max tasks per suite could be added or dragged
 
 0.5.0 / 2018-06-03
 ==================